Create a better way to use potential field plotter
