+++ algorithms/potential_fields/__init__.py
@@ -1,1 +1,2 @@
 import algorithms.potential_fields.fields
+import algorithms.potential_fields.plotter
