+++ strategy/__init__.py
@@ -2,3 +2,7 @@ from strategy import tests
 from strategy import larc2020
 from strategy import iron2021
 from strategy.BaseStrategy import Strategy
+
+
+# debug tools
+from strategy import DebugTools
